Various changements

+++ libraries/ColorIR/ColorIR.cpp
@@ -3,8 +3,8 @@
 /**
    @return 0 se sotto non c'è niente, 1 se c'è uno specchio e 2 se è nero.
 */
-int ColorIR::read() {
-  int media = analogRead(L_CIR);
+byte ColorIR::read() {
+  byte media = analogRead(L_CIR);
   delay(100);
   media += analogRead(R_CIR);
   media >>= 1;
